AERN2.AccuracySG

Description

The type of Cauchy real numbers

Documentation

data AccuracySG

An accuracy specification which includes a soft target "guide" accuracy in addition to the usual string accuracy requirement.

Constructors

AccuracySG

Fields

_acStrict :: Accuracy

_acGuide :: Accuracy

acSG0 :: AccuracySG

accuracySG :: Accuracy -> AccuracySG

bitsS :: Integer -> AccuracySG

bitsSG :: Integer -> Integer -> AccuracySG