AERN2.MPBallWithGlobalPrec

Description

This type is useful for iRRAM-style arrow-generic computation over MPBall with a global precision.

The protocol and type of objects depending on a global parameter

type MPBallWithGlobalPrecP = WithGlobalParamP Precision MPBall

pMPBallWGPrec :: MPBallWithGlobalPrecP

wgprmName :: WithGlobalParamA to prm a -> String

wgprmId :: WithGlobalParamA to prm a -> Maybe (QAId to)

wgprmSources :: WithGlobalParamA to prm a -> [QAId to]

wgprmRename :: (String -> String) -> WithGlobalParamA to prm a -> WithGlobalParamA to prm a

wgprmQuery :: QAArrow to => WithGlobalParamA to prm a -> Maybe (QAId to) -> prm `to` a

Get an approximation of the limit with at least the specified accuracy. (A specialisation of qaMakeQuery for values with global state.)

(?) :: QAArrow to => QA to p -> Q p `to` A p infix 1

An infix synonym of qaMakeQuery with no source

wgprmQueryA :: QAArrow to => Maybe (QAId to) -> (WithGlobalParamA to prm a, prm) `to` a

wgprmListQueryA :: QAArrow to => Maybe (QAId to) -> ([WithGlobalParamA to prm a], prm) `to` [a]

type MPBallWithGlobalPrecA to = WithGlobalParamA to Precision MPBall

type MPBallWithGlobalPrec = MPBallWithGlobalPrecA (->)

newMPBallWGPrec :: QAArrow to => String -> [AnyProtocolQA to] -> ((Maybe (QAId to), Maybe (QAId to)) -> Precision `to` MPBall) -> MPBallWithGlobalPrecA to

newMPBallWGPrecSimple :: QAArrow to => ((Maybe (QAId to), Maybe (QAId to)) -> Precision `to` MPBall) -> MPBallWithGlobalPrecA to

Orphan instances

QAArrow to0 => ConvertibleExactly Int (MPBallWithGlobalPrecA to0)

Methods

safeConvertExactly :: Int -> ConvertResult (MPBallWithGlobalPrecA to0)

QAArrow to0 => ConvertibleExactly Integer (MPBallWithGlobalPrecA to0)

Methods

safeConvertExactly :: Integer -> ConvertResult (MPBallWithGlobalPrecA to0)

QAArrow to0 => ConvertibleExactly Rational (MPBallWithGlobalPrecA to0)

Methods

safeConvertExactly :: Rational -> ConvertResult (MPBallWithGlobalPrecA to0)

QAArrow to0 => ConvertibleExactly Dyadic (MPBallWithGlobalPrecA to0)

Methods

safeConvertExactly :: Dyadic -> ConvertResult (MPBallWithGlobalPrecA to0)

QAArrow to0 => ConvertibleExactly CauchyReal (MPBallWithGlobalPrecA to0)

Methods

safeConvertExactly :: CauchyReal -> ConvertResult (MPBallWithGlobalPrecA to0)