AERN2.QA

Description

Cacheable question-answer protocols

Documentation

class (Show p, Show (Q p), Show (A p)) => QAProtocol p

Associated Types

type Q p

type A p

class QAProtocol p => QAProtocolCacheable p where

Minimal complete definition

newQACache, lookupQACache, updateQACache

Associated Types

type QACache p

Methods

newQACache :: p -> QACache p

lookupQACache :: p -> QACache p -> Q p -> Maybe (A p)

updateQACache :: p -> QACache p -> Q p -> A p -> QACache p

data QA to p

An object we can ask queries about. Queries can be asked in some Arrow to.

Constructors

QA__

Fields

qaName :: String

qaId :: Maybe (QAId to)

qaSources :: [QAId to]

qaProtocol :: p

qaSampleQ :: Q p

qaMakeQuery :: Q p `to` A p

(?) :: QA to p -> Q p `to` A p infix 1

An infix synonym of qaMakeQuery.

data AnyProtocolQA to

Constructors

QAProtocolCacheable p => AnyProtocolQA (QA to p)

addUnsafeMemoisation :: QAProtocolCacheable p => QA (->) p -> QA (->) p

Add caching to pure (->) QA objects via unsafe memoization, inspired by https://hackage.haskell.org/package/ireal-0.2.3/docs/src/Data-Number-IReal-UnsafeMemo.html#unsafeMemo, which, in turn, is inspired by Lennart Augustsson's uglymemo.

class (ArrowChoice to, Eq (QAId to)) => QAArrow to where

A class of Arrows suitable for use in QA objects.

Minimal complete definition

qaRegister, qaMakeQueryGetPromiseA, qaFulfilPromiseA

Associated Types

type QAId to

type QAPromise to :: * -> *

Methods

qaRegister :: QAProtocolCacheable p => QA to p `to` QA to p

Register a QA object, which leads to a change in its query processing mechanism so that, eg, answers can be cached or computations assigned to different threads/processes.

The "sources" component of the incoming QA object can be used to record the dependency graph among QA objects. After registration, the QA object should have its list of dependencies **empty** as the registration has recorded them elsewhere.

qaMakeQueryGetPromiseA :: (QA to p, Q p) `to` QAPromise to (A p)

qaFulfilPromiseA :: QAPromise to a `to` a

qaMakeQueryA :: (QA to p, Q p) `to` A p

qaMakeQueriesA :: [(QA to p, Q p)] `to` [A p]

newQA :: QAProtocolCacheable p => String -> [AnyProtocolQA to] -> p -> Q p -> (Q p `to` A p) -> QA to p

qaMakeQueryOnManyA :: QAArrow to => ([QA to p], Q p) `to` [A p]

(-:-) :: (QAArrow to, QAProtocolCacheable p) => QA to p `to` QA to p infix 0

qaArr :: (QAArrow to, QAProtocolCacheable p) => QA (->) p -> QA to p

Turn a pure QA object into any QAArrow QA object.

type QACachedA = Kleisli QACachedM

data QANetInfo

Constructors

QANetInfo

Fields

net_id2value :: Map ValueId AnyQAComputation

net_log :: QANetLog

executeQACachedA :: (() `QACachedA` a) -> (QANetLog, a)

printQANetLogThenResult :: Show a => (QANetLog, a) -> IO ()