AERN2.Real

Description

Cauchy reals

The type of real numbers

data CauchyRealP

pCR :: CauchyRealP

realName :: CauchyRealA to -> String

realId :: CauchyRealA to -> Maybe (QAId to)

realSources :: CauchyRealA to -> [QAId to]

realRename :: (String -> String) -> CauchyRealA to -> CauchyRealA to

realWithAccuracy :: CauchyRealA to -> Accuracy `to` MPBall

Get a ball approximation of the real number with at least the specified accuracy. (A specialisation of qaMakeQuery for Cauchy reals.)

(?) :: QA to p -> Q p `to` A p infix 1

An infix synonym of qaMakeQuery.

realWithAccuracyA :: QAArrow to => (CauchyRealA to, Accuracy) `to` MPBall

type CauchyRealA to = QA to CauchyRealP

type CauchyReal = CauchyRealA (->)

newCR :: QAArrow to => String -> [AnyProtocolQA to] -> (Accuracy `to` MPBall) -> CauchyRealA to

convergentList2CauchyRealA :: QAArrow to => String -> [MPBall] -> CauchyRealA to

seqByPrecision2CauchyRealA :: QAArrow to => String -> (Precision -> MPBall) -> CauchyRealA to

type CanBeReal t = CanBeRealA (->) t

real :: CanBeRealA (->) t => t -> CauchyReal

type CanBeRealA to t = ConvertibleExactly t (CauchyRealA to)

realA :: CanBeRealA to t => t -> CauchyRealA to

constants

pi :: CauchyReal

To get pi in an arbitrary arrow, use piA.

piA :: QAArrow to => CauchyRealA to

arrow-generic versions of some operations

expA :: (QAArrow to, CanExp t, ExpType t ~ CauchyReal) => t -> CauchyRealA to

logA :: (QAArrow to, CanLog t, LogType t ~ CauchyReal) => t -> CauchyRealA to

sqrtA :: (QAArrow to, CanSqrt t, SqrtType t ~ CauchyReal) => t -> CauchyRealA to

sinA :: (QAArrow to, CanSinCos t, SinCosType t ~ CauchyReal) => t -> CauchyRealA to

cosA :: (QAArrow to, CanSinCos t, SinCosType t ~ CauchyReal) => t -> CauchyRealA to

auxiliary functions for making new real operations

unaryOp :: QAArrow to => String -> (CatchingNumExceptions MPBall -> CatchingNumExceptions MPBall) -> (CauchyRealA to -> Accuracy `to` (Accuracy, Maybe MPBall)) -> CauchyRealA to -> CauchyRealA to

binaryOp :: QAArrow to => String -> (CatchingNumExceptions MPBall -> CatchingNumExceptions MPBall -> CatchingNumExceptions MPBall) -> (CauchyRealA to -> CauchyRealA to -> Accuracy `to` ((Accuracy, Maybe MPBall), (Accuracy, Maybe MPBall))) -> CauchyRealA to -> CauchyRealA to -> CauchyRealA to

binaryOpWithPureArg :: QAArrow to => String -> (CatchingNumExceptions MPBall -> t -> CatchingNumExceptions MPBall) -> (CauchyRealA to -> t -> Accuracy `to` (Accuracy, Maybe MPBall)) -> CauchyRealA to -> t -> CauchyRealA to

getCRFnNormLog :: Arrow to => CauchyRealA to -> (MPBall -> MPBall) -> Accuracy `to` (NormLog, MPBall)

getInitQ1FromSimple :: Arrow to => (Accuracy `to` q) -> r1 -> Accuracy `to` (q, Maybe MPBall)

getInitQ1TFromSimple :: Arrow to => (Accuracy `to` q) -> r1 -> t -> Accuracy `to` (q, Maybe MPBall)

getInitQ1Q2FromSimple :: Arrow to => (Accuracy `to` (q, q)) -> r1 -> r2 -> Accuracy `to` ((q, Maybe MPBall), (q, Maybe MPBall))

binaryWithBall :: (MPBall -> MPBall -> MPBall) -> CauchyReal -> MPBall -> MPBall