AERN2.Real.Arithmetic

Description

Arithmetic operations on Cauchy Real numbers, except those that are defined for more general sequences

Documentation

pi :: CauchyReal

To get pi in an arbitrary arrow, use piA.

piA :: QAArrow to => CauchyRealA to

Orphan instances

Floating CauchyRealCN

Methods

pi :: CauchyRealCN

exp :: CauchyRealCN -> CauchyRealCN

log :: CauchyRealCN -> CauchyRealCN

sqrt :: CauchyRealCN -> CauchyRealCN

(**) :: CauchyRealCN -> CauchyRealCN -> CauchyRealCN

logBase :: CauchyRealCN -> CauchyRealCN -> CauchyRealCN

sin :: CauchyRealCN -> CauchyRealCN

cos :: CauchyRealCN -> CauchyRealCN

tan :: CauchyRealCN -> CauchyRealCN

asin :: CauchyRealCN -> CauchyRealCN

acos :: CauchyRealCN -> CauchyRealCN

atan :: CauchyRealCN -> CauchyRealCN

sinh :: CauchyRealCN -> CauchyRealCN

cosh :: CauchyRealCN -> CauchyRealCN

tanh :: CauchyRealCN -> CauchyRealCN

asinh :: CauchyRealCN -> CauchyRealCN

acosh :: CauchyRealCN -> CauchyRealCN

atanh :: CauchyRealCN -> CauchyRealCN

log1p :: CauchyRealCN -> CauchyRealCN

expm1 :: CauchyRealCN -> CauchyRealCN

log1pexp :: CauchyRealCN -> CauchyRealCN

log1mexp :: CauchyRealCN -> CauchyRealCN

Floating CauchyReal

Methods

pi :: CauchyReal

exp :: CauchyReal -> CauchyReal

log :: CauchyReal -> CauchyReal

sqrt :: CauchyReal -> CauchyReal

(**) :: CauchyReal -> CauchyReal -> CauchyReal

logBase :: CauchyReal -> CauchyReal -> CauchyReal

sin :: CauchyReal -> CauchyReal

cos :: CauchyReal -> CauchyReal

tan :: CauchyReal -> CauchyReal

asin :: CauchyReal -> CauchyReal

acos :: CauchyReal -> CauchyReal

atan :: CauchyReal -> CauchyReal

sinh :: CauchyReal -> CauchyReal

cosh :: CauchyReal -> CauchyReal

tanh :: CauchyReal -> CauchyReal

asinh :: CauchyReal -> CauchyReal

acosh :: CauchyReal -> CauchyReal

atanh :: CauchyReal -> CauchyReal

log1p :: CauchyReal -> CauchyReal

expm1 :: CauchyReal -> CauchyReal

log1pexp :: CauchyReal -> CauchyReal

log1mexp :: CauchyReal -> CauchyReal

CanSqrt Int

Associated Types

type SqrtType Int :: *

Methods

sqrt :: Int -> SqrtType Int

CanSqrt Integer

Associated Types

type SqrtType Integer :: *

Methods

sqrt :: Integer -> SqrtType Integer

CanSqrt Rational

Associated Types

type SqrtType Rational :: *

Methods

sqrt :: Rational -> SqrtType Rational

CanSqrt Dyadic

Associated Types

type SqrtType Dyadic :: *

Methods

sqrt :: Dyadic -> SqrtType Dyadic

CanExp Int

Associated Types

type ExpType Int :: *

Methods

exp :: Int -> ExpType Int

CanExp Integer

Associated Types

type ExpType Integer :: *

Methods

exp :: Integer -> ExpType Integer

CanExp Rational

Associated Types

type ExpType Rational :: *

Methods

exp :: Rational -> ExpType Rational

CanExp Dyadic

Associated Types

type ExpType Dyadic :: *

Methods

exp :: Dyadic -> ExpType Dyadic

CanLog Int

Associated Types

type LogType Int :: *

Methods

log :: Int -> LogType Int

CanLog Integer

Associated Types

type LogType Integer :: *

Methods

log :: Integer -> LogType Integer

CanLog Rational

Associated Types

type LogType Rational :: *

Methods

log :: Rational -> LogType Rational

CanLog Dyadic

Associated Types

type LogType Dyadic :: *

Methods

log :: Dyadic -> LogType Dyadic

CanSinCos Int

Associated Types

type SinCosType Int :: *

Methods

cos :: Int -> SinCosType Int

sin :: Int -> SinCosType Int

CanSinCos Integer

Associated Types

type SinCosType Integer :: *

Methods

cos :: Integer -> SinCosType Integer

sin :: Integer -> SinCosType Integer

CanSinCos Rational

Associated Types

type SinCosType Rational :: *

Methods

cos :: Rational -> SinCosType Rational

sin :: Rational -> SinCosType Rational

CanSinCos Dyadic

Associated Types

type SinCosType Dyadic :: *

Methods

cos :: Dyadic -> SinCosType Dyadic

sin :: Dyadic -> SinCosType Dyadic

Convertible CauchyReal Double

Methods

safeConvert :: CauchyReal -> ConvertResult Double

CanDiv Double CauchyReal

Associated Types

type DivTypeNoCN Double CauchyReal :: *

type DivType Double CauchyReal :: *

Methods

divideNoCN :: Double -> CauchyReal -> DivTypeNoCN Double CauchyReal

divide :: Double -> CauchyReal -> DivType Double CauchyReal

CanDiv CauchyReal Double

Associated Types

type DivTypeNoCN CauchyReal Double :: *

type DivType CauchyReal Double :: *

Methods

divideNoCN :: CauchyReal -> Double -> DivTypeNoCN CauchyReal Double

divide :: CauchyReal -> Double -> DivType CauchyReal Double

CanMulAsymmetric Double CauchyReal

Associated Types

type MulType Double CauchyReal :: *

Methods

mul :: Double -> CauchyReal -> MulType Double CauchyReal

CanMulAsymmetric CauchyReal Double

Associated Types

type MulType CauchyReal Double :: *

Methods

mul :: CauchyReal -> Double -> MulType CauchyReal Double

CanPow Double CauchyReal

Associated Types

type PowTypeNoCN Double CauchyReal :: *

type PowType Double CauchyReal :: *

Methods

powNoCN :: Double -> CauchyReal -> PowTypeNoCN Double CauchyReal

pow :: Double -> CauchyReal -> PowType Double CauchyReal

CanPow Int Rational

Associated Types

type PowTypeNoCN Int Rational :: *

type PowType Int Rational :: *

Methods

powNoCN :: Int -> Rational -> PowTypeNoCN Int Rational

pow :: Int -> Rational -> PowType Int Rational

CanPow Int Dyadic

Associated Types

type PowTypeNoCN Int Dyadic :: *

type PowType Int Dyadic :: *

Methods

powNoCN :: Int -> Dyadic -> PowTypeNoCN Int Dyadic

pow :: Int -> Dyadic -> PowType Int Dyadic

CanPow Integer Rational

Associated Types

type PowTypeNoCN Integer Rational :: *

type PowType Integer Rational :: *

Methods

powNoCN :: Integer -> Rational -> PowTypeNoCN Integer Rational

pow :: Integer -> Rational -> PowType Integer Rational

CanPow Integer Dyadic

Associated Types

type PowTypeNoCN Integer Dyadic :: *

type PowType Integer Dyadic :: *

Methods

powNoCN :: Integer -> Dyadic -> PowTypeNoCN Integer Dyadic

pow :: Integer -> Dyadic -> PowType Integer Dyadic

CanPow Rational Rational

Associated Types

type PowTypeNoCN Rational Rational :: *

type PowType Rational Rational :: *

Methods

powNoCN :: Rational -> Rational -> PowTypeNoCN Rational Rational

pow :: Rational -> Rational -> PowType Rational Rational

CanPow Rational Dyadic

Associated Types

type PowTypeNoCN Rational Dyadic :: *

type PowType Rational Dyadic :: *

Methods

powNoCN :: Rational -> Dyadic -> PowTypeNoCN Rational Dyadic

pow :: Rational -> Dyadic -> PowType Rational Dyadic

CanPow Dyadic Rational

Associated Types

type PowTypeNoCN Dyadic Rational :: *

type PowType Dyadic Rational :: *

Methods

powNoCN :: Dyadic -> Rational -> PowTypeNoCN Dyadic Rational

pow :: Dyadic -> Rational -> PowType Dyadic Rational

CanPow Dyadic Dyadic

Associated Types

type PowTypeNoCN Dyadic Dyadic :: *

type PowType Dyadic Dyadic :: *

Methods

powNoCN :: Dyadic -> Dyadic -> PowTypeNoCN Dyadic Dyadic

pow :: Dyadic -> Dyadic -> PowType Dyadic Dyadic

CanPow CauchyReal Double

Associated Types

type PowTypeNoCN CauchyReal Double :: *

type PowType CauchyReal Double :: *

Methods

powNoCN :: CauchyReal -> Double -> PowTypeNoCN CauchyReal Double

pow :: CauchyReal -> Double -> PowType CauchyReal Double

CanAddAsymmetric Double CauchyReal

Associated Types

type AddType Double CauchyReal :: *

Methods

add :: Double -> CauchyReal -> AddType Double CauchyReal

CanAddAsymmetric CauchyReal Double

Associated Types

type AddType CauchyReal Double :: *

Methods

add :: CauchyReal -> Double -> AddType CauchyReal Double

CanSub Double CauchyReal

Associated Types

type SubType Double CauchyReal :: *

Methods

sub :: Double -> CauchyReal -> SubType Double CauchyReal

CanSub CauchyReal Double

Associated Types

type SubType CauchyReal Double :: *

Methods

sub :: CauchyReal -> Double -> SubType CauchyReal Double

QAArrow to => Field (CauchyRealA to)

QAArrow to => OrderedField (CauchyRealA to)

QAArrow to => Ring (CauchyRealA to)

QAArrow to => OrderedRing (CauchyRealA to)

CanMulAsymmetric t (CauchyRealA to) => CanMulAsymmetric (Complex t) (CauchyRealA to)

Associated Types

type MulType (Complex t) (CauchyRealA to) :: *

Methods

mul :: Complex t -> CauchyRealA to -> MulType (Complex t) (CauchyRealA to)

CanMulAsymmetric (CauchyRealA to) t => CanMulAsymmetric (CauchyRealA to) (Complex t)

Associated Types

type MulType (CauchyRealA to) (Complex t) :: *

Methods

mul :: CauchyRealA to -> Complex t -> MulType (CauchyRealA to) (Complex t)

(QAArrow to, CanAddAsymmetric t (CauchyRealA to)) => CanAddAsymmetric (Complex t) (CauchyRealA to)

Associated Types

type AddType (Complex t) (CauchyRealA to) :: *

Methods

add :: Complex t -> CauchyRealA to -> AddType (Complex t) (CauchyRealA to)

(QAArrow to, CanAddAsymmetric (CauchyRealA to) t) => CanAddAsymmetric (CauchyRealA to) (Complex t)

Associated Types

type AddType (CauchyRealA to) (Complex t) :: *

Methods

add :: CauchyRealA to -> Complex t -> AddType (CauchyRealA to) (Complex t)

(QAArrow to, CanAdd t (CauchyRealA to)) => CanSub (Complex t) (CauchyRealA to)

Associated Types

type SubType (Complex t) (CauchyRealA to) :: *

Methods

sub :: Complex t -> CauchyRealA to -> SubType (Complex t) (CauchyRealA to)

(QAArrow to, CanAdd (CauchyRealA to) t, CanNegSameType t) => CanSub (CauchyRealA to) (Complex t)

Associated Types

type SubType (CauchyRealA to) (Complex t) :: *

Methods

sub :: CauchyRealA to -> Complex t -> SubType (CauchyRealA to) (Complex t)