AERN2.Real.Type

Description

The type of Cauchy real numbers

Documentation

data CauchyRealP

pCR :: CauchyRealP

realName :: CauchyRealA to -> String

realId :: CauchyRealA to -> Maybe (QAId to)

realSources :: CauchyRealA to -> [QAId to]

realRename :: (String -> String) -> CauchyRealA to -> CauchyRealA to

realWithAccuracy :: QAArrow to => CauchyRealA to -> AccuracySG `to` MPBall

Get a ball approximation of the real number with at least the specified accuracy. (A specialisation of qaMakeQuery for Cauchy reals.)

realWithAccuracyA :: QAArrow to => (CauchyRealA to, AccuracySG) `to` MPBall

realsWithAccuracyA :: QAArrow to => ([CauchyRealA to], AccuracySG) `to` [MPBall]

type CauchyRealA to = QA to CauchyRealP

type CauchyReal = CauchyRealA (->)

newCR :: QAArrow to => String -> [AnyProtocolQA to] -> (AccuracySG `to` MPBall) -> CauchyRealA to

convergentList2CauchyRealA :: QAArrow to => String -> [MPBall] -> CauchyRealA to

seqByPrecision2CauchyRealA :: QAArrow to => String -> (Precision -> MPBall) -> CauchyRealA to

type CanBeReal t = CanBeRealA (->) t

real :: CanBeRealA (->) t => t -> CauchyReal

type CanBeRealA to t = ConvertibleExactly t (CauchyRealA to)

realA :: CanBeRealA to t => t -> CauchyRealA to

type CanBeComplex t = CanBeComplexA (->) t

complex :: CanBeComplexA (->) t => t -> Complex CauchyReal

type CanBeComplexA to t = ConvertibleExactly t (Complex (CauchyRealA to))

complexA :: CanBeComplexA to t => t -> Complex (CauchyRealA to)

pickNonZeroRealA :: QAArrow to => [(CauchyRealA to, s)] `to` (CauchyRealA to, s)

Given a list [(a1,b1),(a2,b2),...] and assuming that at least one of a1,a2,... is non-zero, pick one of them and return the corresponding pair (ai,bi).

If none of a1,a2,... is zero, either throw an exception or loop forever.