AERN2.Sequence

Description

A type of fast convergent sequences parametrised by the arrow in which the elements of the sequence are queried

Documentation

module AERN2.AccuracySG

The protocol and type of fast converging sequences

data SequenceP a

Constructors

SequenceP

Fields

unSequenceP :: a

pSeq :: a -> SequenceP a

class (Show a, Show (EnsureNoCN a), Show (EnsureCN a), HasAccuracy a, CanAdjustToAccuracySG a, CanEnsureCN a, CanEnsureCN (EnsureCN a), HasAccuracy (EnsureNoCN a), CanIntersectCNSameType a) => SuitableForSeq a

type SequenceA to a = QA to (SequenceP a)

type Sequence a = SequenceA (->) a

newSeq :: (QAArrow to, SuitableForSeq a) => a -> String -> [AnyProtocolQA to] -> ((Maybe (QAId to), Maybe (QAId to)) -> AccuracySG `to` a) -> SequenceA to a

newSeqSimple :: (QAArrow to, SuitableForSeq a) => a -> ((Maybe (QAId to), Maybe (QAId to)) -> AccuracySG `to` a) -> SequenceA to a

seqName :: SequenceA to a -> String

seqId :: SequenceA to a -> Maybe (QAId to)

seqSources :: SequenceA to a -> [QAId to]

seqRename :: (String -> String) -> SequenceA to a -> SequenceA to a

seqWithAccuracy :: QAArrow to => SequenceA to a -> Maybe (QAId to) -> AccuracySG `to` a

Get an approximation of the limit with at least the specified accuracy. (A specialisation of qaMakeQuery for Cauchy sequences.)

(?) :: QAArrow to => QA to p -> Q p `to` A p infix 1

An infix synonym of qaMakeQuery with no source

seqWithAccuracyA :: QAArrow to => Maybe (QAId to) -> (SequenceA to a, AccuracySG) `to` a

seqsWithAccuracyA :: QAArrow to => Maybe (QAId to) -> ([SequenceA to a], AccuracySG) `to` [a]

(-:-) :: (QAArrow to, QAProtocolCacheable p) => QA to p `to` QA to p infix 0

An infix synonym of qaRegister

(-:-||) :: (QAArrow to, QAProtocolCacheable p) => QA to p `to` QA to p infix 0

An infix synonym of qaRegister

(-:-|) :: (QAArrow to, QAProtocolCacheable p) => QA to p `to` QA to p infix 0

An infix synonym of qaRegister

data SequenceAtAccuracy a

SequenceAtAccuracy exists only so that we can QuickCheck that Sequence satisfies properties whose statement relies on an instance of HasEqCertainly. Sequence is not an instance but SequenceAtAccuracy is.

Constructors

SequenceAtAccuracy (Sequence a) AccuracySG

convergentList2SequenceA :: (QAArrow to, SuitableForSeq a) => String -> [a] -> SequenceA to a

seqByPrecision2SequenceA :: (QAArrow to, SuitableForSeq a) => String -> (Precision -> a) -> SequenceA to a

selecting one of several staged computations

pick :: QAArrow to => Maybe (QAId to) -> [SequenceA to (Maybe a)] `to` a

auxiliary functions for making new sequence operations

unaryOp :: (QAArrow to, SuitableForSeq a, SuitableForSeq b) => String -> (a -> b) -> (Maybe (QAId to) -> SequenceA to a -> AccuracySG `to` (AccuracySG, Maybe a)) -> SequenceA to a -> SequenceA to b

binaryOp :: (QAArrow to, SuitableForSeq a, SuitableForSeq b, SuitableForSeq c) => String -> (a -> b -> c) -> (Maybe (QAId to) -> SequenceA to a -> SequenceA to b -> AccuracySG `to` ((AccuracySG, Maybe a), (AccuracySG, Maybe b))) -> SequenceA to a -> SequenceA to b -> SequenceA to c

binaryOpWithPureArg :: (QAArrow to, SuitableForSeq a, SuitableForSeq b) => String -> (a -> t -> b) -> (Maybe (QAId to) -> SequenceA to a -> t -> AccuracySG `to` (AccuracySG, Maybe a)) -> SequenceA to a -> t -> SequenceA to b

getSeqFnNormLog :: (QAArrow to, CanEnsureCN v, HasNorm (EnsureNoCN v)) => Maybe (QAId to) -> SequenceA to a -> (a -> v) -> AccuracySG `to` (Maybe Integer, a)

getInitQ1FromSimple

Arguments

:: Arrow to

=> (AccuracySG `to` q)

-> Maybe (QAId to)
my id

-> r1

-> AccuracySG `to` (q, Maybe a)

getInitQ1TFromSimple

Arguments

:: Arrow to

=> (AccuracySG `to` q)

-> Maybe (QAId to)
my id

-> r1

-> t

-> AccuracySG `to` (q, Maybe a)

getInitQ1Q2FromSimple

Arguments

:: Arrow to

=> (AccuracySG `to` (q, q))

-> Maybe (QAId to)
my id

-> r1

-> r2

-> AccuracySG `to` ((q, Maybe a), (q, Maybe b))