AERN2.Sequence.Branching

Description

Branching operations for sequences

Documentation

type SeqBoolP = SequenceP (Maybe Bool)

type SeqBoolA to = SequenceA to (Maybe Bool)

type SeqBool = SeqBoolA (->)

pBool :: SeqBoolP

data SequenceAtAccuracy a

SequenceAtAccuracy exists only so that we can QuickCheck that Sequence satisfies properties whose statement relies on an instance of HasEqCertainly. Sequence is not an instance but SequenceAtAccuracy is.

Constructors

SequenceAtAccuracy (Sequence a) AccuracySG

pickNonZeroSeqA :: (QAArrow to, CanPickNonZero a) => Maybe (QAId to) -> [(SequenceA to a, s)] `to` Maybe (SequenceA to a, s)

Given a list [(a1,b1),(a2,b2),...] and assuming that at least one of a1,a2,... is non-zero, pick one of them and return the corresponding pair (ai,bi).

If none of a1,a2,... is zero, either throw an exception or loop forever.

pick :: QAArrow to => Maybe (QAId to) -> [SequenceA to (Maybe a)] `to` a

Orphan instances

CanPickNonZero a => CanPickNonZero (Sequence a)

Methods

pickNonZero :: [(Sequence a, s)] -> Maybe (Sequence a, s)

(QAArrow to, ArrowApply to, HasIfThenElse b t, HasIfThenElse b (to AccuracySG t), (~) * (IfThenElseType b (to AccuracySG t)) (to AccuracySG (IfThenElseType b t)), SuitableForSeq b, SuitableForSeq t, SuitableForSeq (IfThenElseType b t)) => HasIfThenElse (SequenceA to b) (SequenceA to t)
"parallel if"

Associated Types

type IfThenElseType (SequenceA to b) (SequenceA to t) :: *

Methods

ifThenElse :: SequenceA to b -> SequenceA to t -> SequenceA to t -> IfThenElseType (SequenceA to b) (SequenceA to t)