AERN2.Utils.Arrows

Description

Miscellaneous arrow-generic functions

Documentation

mapA :: ArrowChoice to => (t1 `to` t2) -> [t1] `to` [t2]

Apply an arrow morphism on all elements of a list

mapWithIndexA :: ArrowChoice to => (Integer -> t1 `to` t2) -> [t1] `to` [t2]

Apply an arrow morphism on all elements of a list

class CanSwitchArrow to1 to2 where

Minimal complete definition

switchArrow

Methods

switchArrow :: (a `to1` b) -> a `to2` b