AERN2.WithGlobalParam

Description

The protocol and type of objects depending on a global parameter

data WithGlobalParamP prm a

Constructors

WithGlobalParamP

Fields

withGlobalState_s :: Maybe prm

withGlobalState_a :: a

pWGParam :: Maybe prm -> a -> WithGlobalParamP prm a

type SuitableForWGParam prm a = (Show a, Show prm, HasOrderCertainly prm prm)

wgprmName :: WithGlobalParamA to prm a -> String

wgprmId :: WithGlobalParamA to prm a -> Maybe (QAId to)

wgprmSources :: WithGlobalParamA to prm a -> [QAId to]

wgprmRename :: (String -> String) -> WithGlobalParamA to prm a -> WithGlobalParamA to prm a

wgprmQuery :: QAArrow to => WithGlobalParamA to prm a -> Maybe (QAId to) -> prm `to` a

Get an approximation of the limit with at least the specified accuracy. (A specialisation of qaMakeQuery for values with global state.)

(?) :: QAArrow to => QA to p -> Q p `to` A p infix 1

An infix synonym of qaMakeQuery with no source

wgprmQueryA :: QAArrow to => Maybe (QAId to) -> (WithGlobalParamA to prm a, prm) `to` a

wgprmListQueryA :: QAArrow to => Maybe (QAId to) -> ([WithGlobalParamA to prm a], prm) `to` [a]

type WithGlobalParamA to prm a = QA to (WithGlobalParamP prm a)

type WithGlobalParam prm a = WithGlobalParamA (->) prm a

newWGParam :: (QAArrow to, SuitableForWGParam prm a) => Maybe prm -> a -> String -> [AnyProtocolQA to] -> ((Maybe (QAId to), Maybe (QAId to)) -> prm `to` a) -> WithGlobalParamA to prm a

newWGParamSimple :: (QAArrow to, SuitableForWGParam prm a) => Maybe prm -> a -> ((Maybe (QAId to), Maybe (QAId to)) -> prm `to` a) -> WithGlobalParamA to prm a

fmapWGParam :: Arrow to => (a -> b) -> WithGlobalParamA to prm a -> WithGlobalParamA to prm b

auxiliary functions for making new operations

unaryOp :: (QAArrow to, SuitableForWGParam prm a, SuitableForWGParam prm b) => String -> (a -> b) -> WithGlobalParamA to prm a -> WithGlobalParamA to prm b

binaryOp :: (QAArrow to, SuitableForWGParam prm a, SuitableForWGParam prm b, SuitableForWGParam prm c) => String -> (a -> b -> c) -> WithGlobalParamA to prm a -> WithGlobalParamA to prm b -> WithGlobalParamA to prm c

binaryOpWithPureArg :: (QAArrow to, SuitableForWGParam prm a, SuitableForWGParam prm c) => String -> (a -> t -> c) -> WithGlobalParamA to prm a -> t -> WithGlobalParamA to prm c