AERN2.AccuracySG

Description

strict and guide accuracy pairs

Documentation

data AccuracySG

An accuracy specification which includes a soft target "guide" accuracy in addition to the usual string accuracy requirement.

Constructors

AccuracySG

Fields

_acStrict :: Accuracy

_acGuide :: Accuracy

acSG0 :: AccuracySG

default_acSG :: AccuracySG

accuracySG :: Accuracy -> AccuracySG

bitsS :: Integer -> AccuracySG

bitsSG :: Integer -> Integer -> AccuracySG

accuracySGdefaultTolerance :: Integer

class CanAdjustToAccuracySG t where

Minimal complete definition

adjustToAccuracySG

Methods

adjustToAccuracySG :: AccuracySG -> t -> t