AERN2.QA.Protocol

Description

Cacheable question-answer protocols

QA protocols and objects

class (Show p, Show (Q p), Show (A p)) => QAProtocol p

A QA protocol at this level is simply a pair of types.

Associated Types

type Q p

type A p

class (QAProtocol p, HasOrderCertainly (Q p) (Q p)) => QAProtocolCacheable p where

A QA protocol with a caching method.

Minimal complete definition

newQACache, lookupQACache, updateQACache

Associated Types

type QACache p

Methods

newQACache :: p -> QACache p

lookupQACache :: p -> QACache p -> Q p -> (Maybe (A p), Maybe String)

updateQACache :: p -> Q p -> A p -> QACache p -> QACache p

data QA to p

An object we can ask queries about. Queries can be asked in some Arrow to.

Constructors

QA__

Fields

qaName :: String

qaId :: Maybe (QAId to)

qaSources :: [QAId to]

qaProtocol :: p

qaSampleQ :: Maybe (Q p)

qaMakeQueryGetPromise :: (Maybe (QAId to), Maybe (QAId to)) -> Q p `to` QAPromiseA to (A p)

type QAPromiseA to a = () `to` a

(?..) :: QA to p -> Q p `to` QAPromiseA to (A p) infix 1

An infix synonym of qaMakeQuery.

qaRename :: (String -> String) -> QA to p -> QA to p

mapQA :: Arrow to => (p1 -> p2) -> (Q p1 -> Q p2) -> (Q p2 -> Q p1) -> (A p1 -> A p2) -> QA to p1 -> QA to p2

mapQAsameQ :: (Arrow to, Q p1 ~ Q p2) => (p1 -> p2) -> (A p1 -> A p2) -> QA to p1 -> QA to p2

QAArrows

data AnyProtocolQA to

Constructors

QAProtocolCacheable p => AnyProtocolQA (QA to p)

class (ArrowChoice to, Eq (QAId to)) => QAArrow to where

A class of Arrows suitable for use in QA objects.

Minimal complete definition

qaRegister, qaFulfilPromiseA, qaMakeQueryGetPromiseA

Associated Types

type QAId to

Methods

qaRegister :: QAProtocolCacheable p => [QARegOption] -> QA to p `to` QA to p

Register a QA object, which leads to a change in its query processing mechanism so that, eg, answers can be cached or computations assigned to different threads/processes.

The "sources" component of the incoming QA object can be used to record the dependency graph among QA objects. After registration, the QA object should have its list of dependencies **empty** as the registration has recorded them elsewhere.

newQA :: QAProtocolCacheable p => String -> [AnyProtocolQA to] -> p -> Maybe (Q p) -> ((Maybe (QAId to), Maybe (QAId to)) -> Q p `to` A p) -> QA to p

Create a qa object. The object is not "registered" automatically. Invoking this function does not lead to any to-arrow computation. The function is an operation of QAArrow so that for some arrows, the question-answer mechanism can be automatically altered. In particular, this is used to make all objects in the (->) arrow automatically (unsafely) caching their answers. For most arrows, the default implementation is sufficient.

qaFulfilPromiseA :: QAPromiseA to a `to` a

qaMakeQueryGetPromiseA :: Maybe (QAId to) -> (QA to p, Q p) `to` QAPromiseA to (A p)

defaultNewQA :: (QAArrow to, QAProtocolCacheable p) => String -> [AnyProtocolQA to] -> p -> Maybe (Q p) -> ((Maybe (QAId to), Maybe (QAId to)) -> Q p `to` A p) -> QA to p

data QARegOption

Constructors

QARegPreferParallel

QARegPreferSerial

qaMakeQuery

Arguments

:: QAArrow to

=> QA to p

-> Maybe (QAId to)

-> Q p `to` A p
composition of qaMakeQueryGetPromise and the execution of the promise

qaMakeQueryA :: QAArrow to => Maybe (QAId to) -> (QA to p, Q p) `to` A p

qaMakeQueriesA :: QAArrow to => Maybe (QAId to) -> [(QA to p, Q p)] `to` [A p]

qaMakeQueryOnManyA :: QAArrow to => Maybe (QAId to) -> ([QA to p], Q p) `to` [A p]

(?<-) :: QAArrow to => QA to p -> Maybe (QAId to) -> Q p `to` A p infix 1

An infix synonym of qaMakeQuery

(?) :: QAArrow to => QA to p -> Q p `to` A p infix 1

An infix synonym of qaMakeQuery with no source

(-:-) :: (QAArrow to, QAProtocolCacheable p) => QA to p `to` QA to p infix 0

An infix synonym of qaRegister

(-:-|) :: (QAArrow to, QAProtocolCacheable p) => QA to p `to` QA to p infix 0

An infix synonym of qaRegister

(-:-||) :: (QAArrow to, QAProtocolCacheable p) => QA to p `to` QA to p infix 0

An infix synonym of qaRegister

(-?<-) :: QAArrow to => Maybe (QAId to) -> (QA to p, Q p) `to` A p infix 0

An infix synonym of qaMakeQueryA

(-?-) :: QAArrow to => (QA to p, Q p) `to` A p

An infix synonym of qaMakeQueryA with no source

(-?..<-) :: QAArrow to => Maybe (QAId to) -> (QA to p, Q p) `to` QAPromiseA to (A p) infix 0

An infix synonym of qaMakeQueryGetPromiseA

(-?..-) :: QAArrow to => (QA to p, Q p) `to` QAPromiseA to (A p)

An infix synonym of qaMakeQueryGetPromiseA with no source

(-???<-) :: QAArrow to => Maybe (QAId to) -> [(QA to p, Q p)] `to` [A p] infix 0

An infix synonym of qaMakeQueriesA

(-<?<->-) :: QAArrow to => Maybe (QAId to) -> ([QA to p], Q p) `to` [A p] infix 0

An infix synonym of qaMakeQueryOnManyA

qaMake2Queries :: QAArrow to => (QA to p1, QA to p2) -> Maybe (QAId to) -> (Q p1, Q p2) `to` (A p1, A p2)

Run two queries in an interleaving manner, enabling parallelism.

(??<-) :: QAArrow to => (QA to p1, QA to p2) -> Maybe (QAId to) -> (Q p1, Q p2) `to` (A p1, A p2) infix 0

An infix synonym of qaMake2Queries.

qaMake3Queries :: QAArrow to => (QA to p1, QA to p2, QA to p3) -> Maybe (QAId to) -> (Q p1, Q p2, Q p3) `to` (A p1, A p2, A p3)

Run two queries in an interleaving manner, enabling parallelism.