AERN2.Real

Description

Exact real numbers represented by fast-converging Cauchy sequences of MPBalls.

Re-exported dependencies

module AERN2.MP

module AERN2.AccuracySG

The type of real numbers

type CauchyRealP = SequenceP MPBall

pCR :: CauchyRealP

type CauchyRealA to = SequenceA to MPBall

type CauchyReal = CauchyRealA (->)

newCR :: QAArrow to => String -> [AnyProtocolQA to] -> ((Maybe (QAId to), Maybe (QAId to)) -> AccuracySG `to` MPBall) -> CauchyRealA to

type CauchyRealCNA to = SequenceA to (CN MPBall)

type CauchyRealCN = CauchyRealCNA (->)

newCRCN :: QAArrow to => String -> [AnyProtocolQA to] -> ((Maybe (QAId to), Maybe (QAId to)) -> AccuracySG `to` CN MPBall) -> CauchyRealCNA to

Sequence ops specialised to reals

realName :: SequenceA to a -> String

realId :: QA to p -> Maybe (QAId to)

realSources :: QA to p -> [QAId to]

realRename :: (String -> String) -> SequenceA to a -> SequenceA to a

realWithAccuracy :: QAArrow to => CauchyRealA to -> AccuracySG `to` MPBall

Get a ball approximation of the real number with at least the specified accuracy. (A specialisation of qaMakeQuery for Cauchy reals.)

(?) :: QAArrow to => QA to p -> Q p `to` A p infix 1

An infix synonym of qaMakeQuery with no source

realWithAccuracyA :: QAArrow to => Maybe (QAId to) -> (CauchyRealA to, AccuracySG) `to` MPBall

realsWithAccuracyA :: QAArrow to => Maybe (QAId to) -> ([CauchyRealA to], AccuracySG) `to` [MPBall]

(-:-) :: (QAArrow to, QAProtocolCacheable p) => QA to p `to` QA to p infix 0

An infix synonym of qaRegister

convergentList2CauchyRealA :: QAArrow to => String -> [MPBall] -> CauchyRealA to

seqByPrecision2CauchyRealA :: QAArrow to => String -> (Precision -> MPBall) -> CauchyRealA to

Conversions

type CanBeReal t = CanBeRealA (->) t

real :: CanBeRealA (->) t => t -> CauchyReal

type CanBeRealA to t = ConvertibleExactly t (CauchyRealA to)

realA :: CanBeRealA to t => t -> CauchyRealA to

type CanBeComplex t = CanBeComplexA (->) t

complex :: CanBeComplexA (->) t => t -> Complex CauchyReal

type CanBeComplexA to t = ConvertibleExactly t (Complex (CauchyRealA to))

complexA :: CanBeComplexA to t => t -> Complex (CauchyRealA to)

Constants

pi :: CauchyReal

To get pi in an arbitrary arrow, use piA.

piA :: QAArrow to => CauchyRealA to

Mini demos

_addslACachedPrint :: IO ()

_addslAParPrint :: IO ()

_example_pif :: CauchyReal -> CauchyRealCN

_nsection :: Integer -> (Rational -> CauchyReal) -> (Rational, Rational) -> CauchyRealCN